Game update
Added background blur to some menus
Introduced Game result scene (Game Over or Victory)
Extended wave system
Fixed bugs

+++ Unity_game/Assets/Scripts/Buttons/LoginButton.cs
@@ -73,7 +73,11 @@ public class LoginButton : MonoBehaviour, IPointerUpHandler
                     }
                 }
 
-                Directory.Delete("data", true);
+                if (Directory.Exists("data"))
+                {
+                    Directory.Delete("data", true);
+                }
+
                 ZipFile.ExtractToDirectory("data.zip", "./");
                 File.Delete("data.zip");
             }
